Admin privilages

+++ db/migrate/20160713014110_add_to_item.rb
@@ -1,7 +1,7 @@
 class AddToItem < ActiveRecord::Migration
   def change
     add_column :items, :name, :string
-    add_column :items, :price, :decimal
+    add_column :items, :price, :integer
     add_column :items, :active, :boolean
     
   end
